+++ processor/src/test/fixtures/expected/com/example/prop/ImmutablePropTypeArezComponentLikeInterfaceBuilder.java
@@ -1,0 +1,79 @@
+package com.example.prop;
+
+import arez.Disposable;
+import arez.component.Identifiable;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.processing.Generated;
+import org.jetbrains.annotations.Contract;
+import react4j.React;
+import react4j.ReactElement;
+import react4j.ReactNode;
+
+@Generated("react4j.processor.React4jProcessor")
+final class ImmutablePropTypeArezComponentLikeInterfaceBuilder {
+  private ImmutablePropTypeArezComponentLikeInterfaceBuilder() {
+  }
+
+  @Nonnull
+  private static Step1 newBuilder() {
+    return new Builder();
+  }
+
+  @Nonnull
+  @Contract(
+      pure = true
+  )
+  static ReactNode myProp(
+      @Nullable final ImmutablePropTypeArezComponentLikeInterface.MyComponent myProp) {
+    return newBuilder().myProp( myProp );
+  }
+
+  @Nonnull
+  @Contract(
+      pure = true
+  )
+  static ReactNode build() {
+    return newBuilder().build();
+  }
+
+  public interface Step1 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    ReactNode myProp(@Nullable ImmutablePropTypeArezComponentLikeInterface.MyComponent myProp);
+
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    ReactNode build();
+  }
+
+  private static class Builder implements Step1 {
+    @Nonnull
+    private final ReactElement _element = ReactElement.createViewElement( React4j_ImmutablePropTypeArezComponentLikeInterface.Factory.TYPE );
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode myProp(
+        @Nullable final ImmutablePropTypeArezComponentLikeInterface.MyComponent myProp) {
+      _element.setKey( Identifiable.<Object>getArezId( myProp ) + ( React.enableViewNames() ? "_ImmutablePropTypeArezComponentLikeInterface_c528e47c" : ImmutablePropTypeArezComponentLikeInterface.class.getName() ) );
+      assert Disposable.isNotDisposed( myProp );
+      _element.input( React4j_ImmutablePropTypeArezComponentLikeInterface.Inputs.myProp, myProp );
+      return build();
+    }
+
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode build() {
+      return _element;
+    }
+  }
+}