+++ processor/src/test/fixtures/expected/com/example/prop/React4j_MutableDisposableNullableArezComponentTrackingProp.java
@@ -110,7 +110,7 @@ abstract class React4j_MutableDisposableNullableArezComponentTrackingProp extend
     $$react4j$$_state = ViewState.IDLE;
     assert Disposable.isNotDisposed( this );
     final MutableDisposableNullableArezComponentTrackingProp.Model $$react4jv$$_getModel = getModel();
-    if ( null != $$react4jv$$_getModel && !ComponentObservable.observe( $$react4jv$$_getModel ) ) {
+    if ( !ComponentObservable.maybeObserve( $$react4jv$$_getModel ) ) {
       return null;
     }
     if ( Disposable.isDisposed( $$react4jv$$_getModel ) ) {