+++ processor/src/test/fixtures/input/com/example/prop/ImmutablePropTypeArezComponentLikeAndKeyed.java
@@ -1,0 +1,31 @@
+package com.example.prop;
+
+import arez.annotations.ArezComponentLike;
+import javax.annotation.Nullable;
+import react4j.Keyed;
+import react4j.ReactNode;
+import react4j.annotations.Input;
+import react4j.annotations.Render;
+import react4j.annotations.View;
+
+@View
+abstract class ImmutablePropTypeArezComponentLikeAndKeyed
+{
+  @ArezComponentLike
+  interface MyComponent
+    extends Keyed
+  {
+  }
+
+  // Keyed should win over ArezId for immutable prop
+  ImmutablePropTypeArezComponentLikeAndKeyed( @Nullable @Input final MyComponent myProp )
+  {
+  }
+
+  @Nullable
+  @Render
+  ReactNode render()
+  {
+    return null;
+  }
+}