+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Update the `org.realityforge.arez` artifacts to version `0.234`.
+
 ### [v0.212](https://github.com/react4j/react4j/tree/v0.212) (2026-04-15) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.211...v0.212)
 
 Changes in this release: